+++ rpi-led-matrix/src/canvas.rs
@@ -18,6 +18,7 @@ pub struct LedCanvas {
 
 impl LedCanvas {
     /// Retrieves the width & height of the canvas
+    #[must_use]
     pub fn canvas_size(&self) -> (i32, i32) {
         let (mut width, mut height): (c_int, c_int) = (0, 0);
         unsafe {
@@ -40,7 +41,7 @@ impl LedCanvas {
                 color.red,
                 color.green,
                 color.blue,
-            )
+            );
         }
     }
 
@@ -95,6 +96,10 @@ impl LedCanvas {
 
     #[allow(clippy::too_many_arguments)]
     /// Renders text using the C++ library.
+    ///
+    /// # Panics
+    /// If the given `text` fails to convert to a `CString`. This can
+    /// occur when there is a null character mid way in the string.
     pub fn draw_text(
         &mut self,
         font: &LedFont,
@@ -105,7 +110,7 @@ impl LedCanvas {
         kerning_offset: i32,
         vertical: bool,
     ) -> i32 {
-        let ctext = CString::new(text).unwrap();
+        let text = CString::new(text).expect("given string failed to convert into a CString");
         unsafe {
             if vertical {
                 ffi::vertical_draw_text(
@@ -116,7 +121,7 @@ impl LedCanvas {
                     color.red,
                     color.green,
                     color.blue,
-                    ctext.as_ptr(),
+                    text.as_ptr(),
                     kerning_offset as c_int,
                 ) as i32
             } else {
@@ -128,7 +133,7 @@ impl LedCanvas {
                     color.red,
                     color.green,
                     color.blue,
-                    ctext.as_ptr(),
+                    text.as_ptr(),
                     kerning_offset as c_int,
                 ) as i32
             }
@@ -245,7 +250,7 @@ mod tests {
         color.red = 0;
         canvas.fill(&color);
         thread::sleep(time::Duration::new(0, 500000000));
-        matrix.swap(canvas);
+        let _ = matrix.swap(canvas);
         thread::sleep(time::Duration::new(0, 500000000));
     }
 }