+++ rpi-led-matrix-sys/CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - ???
+## UNRELEASED
+
+- Took a lot of clippy suggestions. Added `#[must_use]` where it made sense.
+  Most of the reset was documentation changes and moving from `unwrap` to `expect`.
+
+## [0.1.4] - 2021-12-05
+
+- update C++ submodule to top of tree
+
+## [0.1.3] - 2021-12-05
+
+- Build script improvements around the C++ git submodule
+
+## [0.1.2] - 2021-12-04
+
+- Fix C++ std lib linking issue
+
+## [0.1.1] - 2021-12-04 (**YANKED**)
+
+- Cosmetic README/documentation updates
+
+## [0.1.0] - 2021-12-04 (**YANKED**)
 
 - Initial release